Refactoring the interactors
